+++ src/VocabularyCoach/ViewModels/Interfaces/IEditVocabularyViewModel.cs
@@ -1,8 +1,10 @@
+using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Input;
 using VocabularyCoach.Models;
+using VocabularyCoach.ViewModels.ContextMenu;
 
 namespace VocabularyCoach.ViewModels.Interfaces
 {
@@ -14,11 +16,15 @@ namespace VocabularyCoach.ViewModels.Interfaces
 
 		ObservableCollection<TranslationViewModel> Translations { get; }
 
+		TranslationViewModel SelectedTranslation { get; set; }
+
 		public ICommand SaveChangesCommand { get; }
 
 		public ICommand ClearChangesCommand { get; }
 
 		ICommand GoToStartPageCommand { get; }
+
+		IEnumerable<ContextMenuItem> ContextMenuItems { get; }
 
 		Task Load(Language studiedLanguage, Language knownLanguage, CancellationToken cancellationToken);
 	}