+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added display of progress for the current practice session.
 - Extended start page with vocabulary statistics.
 - Added support of synonyms.
+- Added removal of texts and translations.
 
 ## 1.0.0 - 2023-07-12
 - Initial version.